# backend/app/schemas/auth.py
import uuid
from datetime import datetime

from pydantic import BaseModel, Field

from app.models.user import UserRole


class RegisterBusinessRequest(BaseModel):
    business_name: str
    sector: str | None = None
    owner_full_name: str
    owner_phone: str | None = None
    pin: str = Field(min_length=4, max_length=8)


class LoginRequest(BaseModel):
    business_code: str
    pin: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    user_id: uuid.UUID
    business_id: uuid.UUID
    business_code: str
    business_name: str
    role: UserRole
    full_name: str
    can_view_purchase_prices: bool
    can_view_owner_dashboard: bool
    deletion_scheduled_for: datetime | None = None
    """Set when the owner has asked to delete the business: the apps then show only the
    « suppression en cours » screen (download the history, cancel)."""


class VerifyPinRequest(BaseModel):
    pin: str


class RecoverCodeRequest(BaseModel):
    """Récupération du code entreprise, sans être connecté (nouvel appareil, code perdu).
    Le code seul ne permet aucune action — la connexion exige toujours le PIN — donc une
    vérification d'identité légère suffit : nom de l'entreprise + nom du propriétaire +
    téléphone du propriétaire, tous les trois obligatoires."""

    business_name: str
    owner_full_name: str
    owner_phone: str


class RecoverCodeResponse(BaseModel):
    business_code: str
    business_name: str


class CreateEmployeeRequest(BaseModel):
    full_name: str
    phone: str | None = None
    pin: str = Field(min_length=4, max_length=8)
    can_view_purchase_prices: bool = False
    can_view_owner_dashboard: bool = False


class UpdateEmployeeRequest(BaseModel):
    """Édition/désactivation d'un employé. Pas de `role` (immuable — voir
    auth_service.update_employee) et pas de `pin` (le PIN employé est créé une
    fois à sa création, on ne le réédite pas dans le MVP)."""

    can_view_purchase_prices: bool | None = None
    can_view_owner_dashboard: bool | None = None
    is_active: bool | None = None


class UserOut(BaseModel):
    id: uuid.UUID
    business_id: uuid.UUID
    full_name: str
    phone: str | None
    role: UserRole
    can_view_purchase_prices: bool
    can_view_owner_dashboard: bool
    is_active: bool