# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Korah Business Manager API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/register-business": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_business_post",
        "parameters": [
          {
            "name": "x-device-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          },
          {
            "name": "x-platform",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform"
            }
          },
          {
            "name": "x-app-version",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBusinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Route",
        "operationId": "login_route_auth_login_post",
        "parameters": [
          {
            "name": "x-device-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          },
          {
            "name": "x-platform",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform"
            }
          },
          {
            "name": "x-app-version",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/recover-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Recover Code Route",
        "description": "Code entreprise oublié — sans authentification (c'est tout l'objet).\nVérification d'identité légère : le code seul ne permet aucune action sans le PIN.",
        "operationId": "recover_code_route_auth_recover_code_post",
        "parameters": [
          {
            "name": "x-device-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          },
          {
            "name": "x-platform",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform"
            }
          },
          {
            "name": "x-app-version",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoverCodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/employees": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Employees Route",
        "operationId": "list_employees_route_auth_employees_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "type": "array",
                  "title": "Response List Employees Route Auth Employees Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Employee Route",
        "operationId": "create_employee_route_auth_employees_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/employees/{user_id}": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update Employee Route",
        "description": "Modifier un employé. Le rôle est immuable — un employé ne peut jamais\ndevenir owner via cette route, et aucun employé ne peut se modifier lui-même.",
        "operationId": "update_employee_route_auth_employees__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Delete Employee Route",
        "description": "Supprimer un employé : connexion fermée définitivement, historique conservé.",
        "operationId": "delete_employee_route_auth_employees__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-pin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Pin Route",
        "description": "Re-vérifie le PIN de l'utilisateur connecté avant une action irréversible.",
        "operationId": "verify_pin_route_auth_verify_pin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/admin/auth/login": {
      "post": {
        "tags": [
          "admin-auth"
        ],
        "summary": "Login",
        "operationId": "login_admin_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "AdminLoginRequest"
      },
...
      "AdminTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "super_admin_id": {
            "type": "string",
            "format": "uuid",
            "title": "Super Admin Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/SuperAdminRole"
          },
          "environment": {
            "type": "string",
            "title": "Environment",
            "default": "development"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "super_admin_id",
          "email",
          "full_name",
          "role"
        ],
        "title": "AdminTokenResponse"
      },
...
      "CreateEmployeeRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "pin": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Pin"
          },
          "can_view_purchase_prices": {
            "type": "boolean",
            "title": "Can View Purchase Prices",
            "default": false
          },
          "can_view_owner_dashboard": {
            "type": "boolean",
            "title": "Can View Owner Dashboard",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "pin"
        ],
        "title": "CreateEmployeeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "business_code": {
            "type": "string",
            "title": "Business Code"
          },
          "pin": {
            "type": "string",
            "title": "Pin"
          }
        },
        "type": "object",
        "required": [
          "business_code",
          "pin"
        ],
        "title": "LoginRequest"
      },
...
      "RecoverCodeRequest": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "owner_full_name": {
            "type": "string",
            "title": "Owner Full Name"
          },
          "owner_phone": {
            "type": "string",
            "title": "Owner Phone"
          }
        },
        "type": "object",
        "required": [
          "business_name",
          "owner_full_name",
          "owner_phone"
        ],
        "title": "RecoverCodeRequest",
        "description": "Récupération du code entreprise, sans être connecté (nouvel appareil, code perdu).\nLe code seul ne permet aucune action — la connexion exige toujours le PIN — donc une\nvérification d'identité légère suffit : nom de l'entreprise + nom du propriétaire +\ntéléphone du propriétaire, tous les trois obligatoires."
      },
      "RecoverCodeResponse": {
        "properties": {
          "business_code": {
            "type": "string",
            "title": "Business Code"
          },
          "business_name": {
            "type": "string",
            "title": "Business Name"
          }
        },
        "type": "object",
        "required": [
          "business_code",
          "business_name"
        ],
        "title": "RecoverCodeResponse"
      },
      "RegisterBusinessRequest": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "owner_full_name": {
            "type": "string",
            "title": "Owner Full Name"
          },
          "owner_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Phone"
          },
          "pin": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Pin"
          }
        },
        "type": "object",
        "required": [
          "business_name",
          "owner_full_name",
          "pin"
        ],
        "title": "RegisterBusinessRequest"
      },
...
      "SuperAdminRole": {
        "type": "string",
        "enum": [
          "owner",
          "product_manager",
          "support",
          "developer"
        ],
        "title": "SuperAdminRole"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "business_id": {
            "type": "string",
            "format": "uuid",
            "title": "Business Id"
          },
          "business_code": {
            "type": "string",
            "title": "Business Code"
          },
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "can_view_purchase_prices": {
            "type": "boolean",
            "title": "Can View Purchase Prices"
          },
          "can_view_owner_dashboard": {
            "type": "boolean",
            "title": "Can View Owner Dashboard"
          },
          "deletion_scheduled_for": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletion Scheduled For"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "business_id",
          "business_code",
          "business_name",
          "role",
          "full_name",
          "can_view_purchase_prices",
          "can_view_owner_dashboard"
        ],
        "title": "TokenResponse"
      },
...
      "UpdateEmployeeRequest": {
        "properties": {
          "can_view_purchase_prices": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can View Purchase Prices"
          },
          "can_view_owner_dashboard": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can View Owner Dashboard"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UpdateEmployeeRequest",
        "description": "Édition/désactivation d'un employé. Pas de `role` (immuable — voir\nauth_service.update_employee) et pas de `pin` (le PIN employé est créé une\nfois à sa création, on ne le réédite pas dans le MVP)."
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "business_id": {
            "type": "string",
            "format": "uuid",
            "title": "Business Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "can_view_purchase_prices": {
            "type": "boolean",
            "title": "Can View Purchase Prices"
          },
          "can_view_owner_dashboard": {
            "type": "boolean",
            "title": "Can View Owner Dashboard"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "business_id",
          "full_name",
          "phone",
          "role",
          "can_view_purchase_prices",
          "can_view_owner_dashboard",
          "is_active"
        ],
        "title": "UserOut"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "owner",
          "employee"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyPinRequest": {
        "properties": {
          "pin": {
            "type": "string",
            "title": "Pin"
          }
        },
        "type": "object",
        "required": [
          "pin"
        ],
        "title": "VerifyPinRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}